+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## 3.2.0
 
+- **Feature:**
+  - Add deprecation warnings for `createBitcoinTransaction`, `createEthereumTransaction`, and `getPublicBlockchainAddresses`
+  - Add support for new interchain management endpoints with client functions:
+    - `createBitcoinInterchain`
+    - `createEthereumInterchain`
+    - `updateBitcionInterchain`
+    - `updateEthereumInterchain`
+    - `getInterchainNetwork`
+    - `deleteInterchainNetwork`
+    - `listInterchainNetworks`
+    - `setDefaultInterchainNetwork`
+    - `getDefaultInterchainNetwork`
+    - `signBitcoinTransaction`
+    - `signEthereumTransaction`
 - **Development:**
   - Update development dependencies